Add normalizeDate util and tests
